Refactors main
Generates GMap index only once instead of once per sample

Specifies default /tmp dir for mikado (TODO check if this belongs to the engine configuration instead)
