+++ validation/transcriptome.schema.json
